+++ ApplicationCore/Repositories/ICompanyRepository.cs
@@ -4,7 +4,7 @@ namespace ApplicationCore.Repositories
 {
     public interface ICompanyRepository : IRepository<Company>
     {
-        ValueTask<Company> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+        ValueTask<Company> GetByIdAsync(long id, CancellationToken cancellationToken = default);
         ValueTask<IEnumerable<Company>> GetByFilterAsync(Company filter, CancellationToken cancellationToken = default);
     }
 }
