+++ ApplicationCore/Repositories/IContactRepository.cs
@@ -4,6 +4,6 @@ namespace ApplicationCore.Repositories
 {
     public interface IContactRepository : IRepository<Contact>
     {
-        ValueTask<Contact> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+        ValueTask<Contact> GetByIdAsync(long id, CancellationToken cancellationToken = default);
     }
 }
