+++ ApplicationCore/Services/ICompanyService.cs
@@ -4,10 +4,10 @@ namespace ApplicationCore.Services
 {
     public interface ICompanyService : IBaseService<Company>
     {
-        ValueTask<IEnumerable<Company>> GetByIdsAsync(IEnumerable<int> ids, CancellationToken cancellationToken = default);
+        ValueTask<IEnumerable<Company>> GetByIdsAsync(IEnumerable<long> ids, CancellationToken cancellationToken = default);
         ValueTask<IEnumerable<Company>> GetByFilterAsync(Company filter, CancellationToken cancellationToken = default);
         ValueTask<(IEnumerable<Company> companies, string ids)> CreateCompanyCollection(IEnumerable<Company> companyCollection, CancellationToken cancellationToken = default);
         ValueTask<Company> CreateCompanyAsync(Company company, CancellationToken cancellationToken = default);
-        ValueTask<Company> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+        ValueTask<Company> GetByIdAsync(long id, CancellationToken cancellationToken = default);
     }
 }
