+++ Infrastructure/Data/DbRepository.cs
@@ -29,12 +29,12 @@ namespace Infrastructure.Data
         public async ValueTask<IEnumerable<T>> FindByConditionAsync(Expression<Func<T, bool>> expression, CancellationToken cancellationToken = default)
             => await _dbSet.Where(expression).AsNoTracking().ToListAsync();
 
-        public async ValueTask<T> FindByIdAsync(int id, CancellationToken cancellationToken = default) => await _dbSet.FindAsync(new object[] { id }, cancellationToken);
+        public async ValueTask<T> FindByIdAsync(long id, CancellationToken cancellationToken = default) => await _dbSet.FindAsync(new object[] { id }, cancellationToken);
 
         public async ValueTask<IEnumerable<T>> GetAllAsync(CancellationToken cancellationToken = default) =>
             await _dbSet.ToListAsync(cancellationToken);
         public void Remove(T entity) => _dbSet.Remove(entity);
-        public void RemoveById(int id) => Remove(_dbSet.Find(id));
+        public void RemoveById(long id) => Remove(_dbSet.Find(id));
         public virtual void Update(T entity) => DbContext.Entry(entity).State = EntityState.Modified;
 
         public async ValueTask RollbackAsync(CancellationToken cancellationToken = default)
